requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-redis==5.4.0
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.1
psycopg==3.1.18
django-cors-headers==4.3.1
django-filter==24.1
# Pillow is optional for Python 3.13
Pillow==10.0.0; python_version < "3.13"
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.34
python-dateutil==2.8.2
dj-rest-auth==5.0.2
django-allauth==0.61.1
social-auth-app-django==5.4.0 